Adding Buttons
just started to add buttons. Look up on google "Windows 8 calculator"
and then click on images and its the third image of a scientific
calculator. I have to go but ill continue working on it at 8 when I get
back, so if you want to work on it until then, go ahead.